Introducing DynOS packs
I will also be hosting DynOS packs on this repo now. I will be working on a rework of the README as well soon and have included "anims.zip" in mods/mario-style/anims.zip in case you want to give your custom character model the ability to also do a funny dance.